Implement auto-brightness using exposure
On my camera (Logitech C120) setting the v4l2 brightness control just
affects a digital gain, but does not affect exposure time. It also has
poor built-in auto-exposure, which oversaturates in sunlight.

Now, when auto_bright is enabled, EXPOSURE_AUTO is set to MANUAL and the
EXPOSURE_ABSOLUTE is set based on the computed adjustment. Note that at
least on this camera, EXPOSURE_ABSOLUTE seems to be inverted: a larger
value gives a shorter exposure/dimmer image.